+++ src/main/java/com/simibubi/create/compat/jei/category/BasinCategory.java
@@ -5,22 +5,17 @@ import java.util.List;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 
-import com.simibubi.create.content.processing.recipe.IHeatCondition;
+import com.simibubi.create.api.recipe.HeatCondition;
 
 import net.minecraft.network.chat.Component;
 
 import org.apache.commons.lang3.mutable.MutableInt;
 
-import com.simibubi.create.AllBlocks;
-import com.simibubi.create.AllItems;
 import com.simibubi.create.content.processing.basin.BasinRecipe;
-import com.simibubi.create.content.processing.burner.BlazeBurnerBlock.HeatLevel;
-import com.simibubi.create.content.processing.recipe.HeatCondition;
 import com.simibubi.create.content.processing.recipe.ProcessingOutput;
 import com.simibubi.create.foundation.fluid.FluidIngredient;
 import com.simibubi.create.foundation.gui.AllGuiTextures;
 import com.simibubi.create.foundation.item.ItemHelper;
-import com.simibubi.create.foundation.utility.CreateLang;
 
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
@@ -95,7 +90,7 @@ public class BasinCategory extends CreateRecipeCategory<BasinRecipe> {
 			i++;
 		}
 
-		IHeatCondition heatCondition = recipe.getRequiredHeat();
+		HeatCondition heatCondition = recipe.getRequiredHeat();
 		if (heatCondition == null) return;
 
 		List<ItemStack> itemHints = heatCondition.getItemHints();
@@ -107,7 +102,7 @@ public class BasinCategory extends CreateRecipeCategory<BasinRecipe> {
 
 	@Override
 	public void draw(BasinRecipe recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics graphics, double mouseX, double mouseY) {
-		IHeatCondition heatCondition = recipe.getRequiredHeat();
+		HeatCondition heatCondition = recipe.getRequiredHeat();
 
 		boolean noHeat = heatCondition == null;
