+++ src/main/java/com/simibubi/create/foundation/data/recipe/CreateMixingRecipeGen.java
@@ -4,11 +4,9 @@ import com.simibubi.create.AllFluids;
 import com.simibubi.create.AllItems;
 import com.simibubi.create.AllTags.AllItemTags;
 import com.simibubi.create.Create;
-import com.simibubi.create.api.data.recipe.CompactingRecipeGen;
 import com.simibubi.create.api.data.recipe.MixingRecipeGen;
 
-import com.simibubi.create.api.data.recipe.SequencedAssemblyRecipeGen;
-import com.simibubi.create.content.processing.recipe.HeatCondition;
+import com.simibubi.create.content.processing.recipe.BlazeBurnerHeatCondition;
 import com.simibubi.create.foundation.recipe.BlockTagIngredient;
 
 import net.minecraft.data.PackOutput;
@@ -31,27 +29,27 @@ public final class CreateMixingRecipeGen extends MixingRecipeGen {
 
 	TEMP_LAVA = create("lava_from_cobble", b -> b.require(Tags.Items.COBBLESTONE)
 		.output(Fluids.LAVA, 50)
-		.requiresHeat(HeatCondition.SUPERHEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.SUPERHEATED)),
 
 	TEA = create("tea", b -> b.require(Fluids.WATER, 250)
 		.require(Tags.Fluids.MILK, 250)
 		.require(ItemTags.LEAVES)
 		.output(AllFluids.TEA.get(), 500)
-		.requiresHeat(HeatCondition.HEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.HEATED)),
 
 	CHOCOLATE = create("chocolate", b -> b.require(Tags.Fluids.MILK, 250)
 		.require(Items.SUGAR)
 		.require(Items.COCOA_BEANS)
 		.output(AllFluids.CHOCOLATE.get(), 250)
-		.requiresHeat(HeatCondition.HEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.HEATED)),
 
 	CHOCOLATE_MELTING = create("chocolate_melting", b -> b.require(AllItems.BAR_OF_CHOCOLATE.get())
 		.output(AllFluids.CHOCOLATE.get(), 250)
-		.requiresHeat(HeatCondition.HEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.HEATED)),
 
 	HONEY = create("honey", b -> b.require(Items.HONEY_BLOCK)
 		.output(AllFluids.HONEY.get(), 1000)
-		.requiresHeat(HeatCondition.HEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.HEATED)),
 
 	DOUGH = create("dough_by_mixing", b -> b.require(CreateRecipeProvider.I.wheatFlour())
 		.require(Fluids.WATER, 1000)
@@ -60,7 +58,7 @@ public final class CreateMixingRecipeGen extends MixingRecipeGen {
 	BRASS_INGOT = create("brass_ingot", b -> b.require(CreateRecipeProvider.I.copper())
 		.require(CreateRecipeProvider.I.zinc())
 		.output(AllItems.BRASS_INGOT.get(), 2)
-		.requiresHeat(HeatCondition.HEATED)),
+		.requiresHeat(BlazeBurnerHeatCondition.HEATED)),
 
 	ANDESITE_ALLOY = create("andesite_alloy", b -> b.require(Blocks.ANDESITE)
 		.require(CreateRecipeProvider.I.ironNugget())
